修改用户删除和查找方法的参数类型

+++ src/main/java/com/aidansu/demo/dao/UserDAO.java
@@ -13,9 +13,9 @@ public interface UserDAO {
 
     void update(User user);
 
-    void delete(int id);
+    void delete(long id);
 
-    User findById(int id);
+    User findById(long id);
 
     User findByUsername(String username);
